pyproject.toml
[tool.poetry]
name = "overlore"
version = "0.0.1"
description = "power to the autonomous, generative agency peoples"
authors = ["human spirit <fhuman@spirit.life>"]
repository = "https://github.com/elonmusk/overlore"
documentation = "https://elonmusk.github.io/overlore/"
readme = "README.md"
packages = [{ include = "overlore" }]

[tool.poetry.scripts]
lore-machine = "overlore.main:main"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
python-dotenv = "^1.0.0"
gql = "^3.4.1"
sqlean-py = "^0.21.8.5"
sqlite-vss = "^0.1.2"
responses = "^0.25.0"
starknet-py = "^0.20.0"
respx = "^0.20.2"
openai = "^1.13.3"
guardrails-ai = { git = "https://github.com/guardrails-ai/guardrails.git" }
jsonrpcserver = "^5.0.9"
aiohttp = "^3.9.3"
aiohttp-cors = "^0.7.0"
websockets = "^12.0"
backoff = "^2.2.1"
pydantic = "^2.6.4"
rich = "^13.7.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
deptry = "^0.12.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"
websocket-client = "^1.7.0"
pytest-asyncio = "^0.23.2"
grpcio-tools = "^1.59.3"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.2.7"
mkdocstrings = { extras = ["python"], version = "^0.23.0" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py37']
preview = true

[tool.mypy]
files = ["overlore"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[[tool.mypy.overrides]]
module = "overlore.dojo_world"
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py37"
line-length = 120
fix = true
select = [
	# flake8-2020
	"YTT",
	# flake8-bandit
	"S",
	# flake8-bugbear
	"B",
	# flake8-builtins
	"A",
	# flake8-comprehensions
	"C4",
	# flake8-debugger
	"T10",
	# flake8-simplify
	"SIM",
	# isort
	"I",
	# mccabe
	"C90",
	# pycodestyle
	"E",
	"W",
	# pyflakes
	"F",
	# pygrep-hooks
	"PGH",
	# pyupgrade
	"UP",
	# ruff
	"RUF",
	# tryceratops
	"TRY",
]
ignore = [
	# LineTooLong
	"E501",
	# DoNotAssignLambda
	"E731",
]

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["overlore"]


[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
